+++ CHANGELOG.md
@@ -88,3 +88,6 @@
 - `0.5.12` Start CLI binary uploads to S3.
 - `0.5.13` Add Cursor Plugins and Opencode support. Bug fix in skill serialization. Agent guard installer detects unauthorized modifications of the hooks script. Agent guard installer installs hooks only if requested agents are already installed on the machine.
 - `0.5.14` Fix --insecure usage with the Snyk CLI. Remove unused API calls.
+- `0.5.15` Fix false error triggering on skipping guard installations on missing agents.
+- `0.5.16` Security fixes. Increase analysis request timeout. Add CLI reference docs and refresh JSON output docs.
+- `0.5.17` Introduce support for async analysis. Security fixes related to dependencies.